Silenced fabsf truncation warnings by replacing it with fabs

+++ cocos2d/CCNodeColor.m
@@ -212,7 +212,7 @@
 
 	// Compressed Interpolation mode
 	if( _compressedInterpolation ) {
-		float h2 = 1 / ( fabsf(u.x) + fabsf(u.y) );
+		float h2 = 1 / ( fabs(u.x) + fabs(u.y) );
 		u = ccpMult(u, h2 * (float) sqrt2);
 	}
